+++ SmartLinkV3/src/com/alcatel/smartlinkv3/business/model/UsageSettingModel.java
@@ -9,8 +9,8 @@ import com.alcatel.smartlinkv3.business.statistics.UsageSettingsResult;
 public class UsageSettingModel extends BaseResult{
 	
 	public int HBillingDay = 0;//Billing day (Min:1, Max:31),
-	public int HMonthlyPlan = 0;//The max data that on month could use..
-	public int HUsedData = 0;//The used data in month.
+	public long HMonthlyPlan = 0;//The max data that on month could use..
+	public long HUsedData = 0;//The used data in month.
 	public int HTimeLimitTimes = 0;//The time limit function open, must set the limit time.
 	public int HUsedTimes = 0;//The used time that after open time limit function.
 	public ENUM.OVER_TIME_STATE HTimeLimitFlag = ENUM.OVER_TIME_STATE.Disable;  //The time limit function flage,0: disable 1:enable
